fix decompression of int images
This fixes reading of a compressed lsst.afw.image.MaskedImage.

This is a combination of two fixes: one from me to fix the immediate
cause and one from Bill Pence to fix the root cause.

The immediate cause was from assuming that quantize_level == NO_QUANTIZE
means that you're decompressing a floating-point image.

Bill Pence writes:

    This bug would only show up in the relatively rare case where
    a floating point image that is losslessly compressed with gzip
    is followed by a compressed integer image.

(which is exactly what we do for MaskedImage) and:

    the root cause of this bug is that one of the compression
    parameters for the first floating-point image was not being
    reinitialized when the next integer image was read.  This caused
    the bogus value for the quantize_level.